backend-python/requirements.txt
Flask
Flask-Cors
requests
mysql-connector-python
kubernetes
scikit-learn
pandas
numpy
